Split component into small components. Place them on one lvl. Replace a part of logic from this.state to mobX store (votesStore)
