+++ web/src/components/Vote/Vote.scss
@@ -1,0 +1,39 @@
+.vote-page {
+    h1 {
+        padding-bottom: 1rem;
+        border-bottom: 1px solid rgba(0, 0, 0, 0.12);
+        margin-bottom: 2rem;
+    }
+
+    .candidate-item {
+        display: flex;
+        margin: 2rem 0;
+        align-items: center;
+
+        .votes-count {
+            width: 6rem;
+            display: flex;
+            align-items: center;
+
+            .count {
+                background-color: rgba(0, 0, 0, 0.3);
+                padding: .4rem 1rem;
+                border-radius: .3rem;
+                color: #fff;
+            }
+        }
+
+        .candidate-name {
+            flex: 1;
+        }
+
+        .vote-btn {
+            margin-left: 2rem;
+            cursor: pointer;
+        }
+    }
+
+    .is-voted {
+        font-weight: bold;
+    }
+}
