--- web/src/pages/VotePage/VotePage.scss
@@ -1,88 +1,0 @@
-.vote-page {
-    h1 {
-        padding-bottom: 1rem;
-        border-bottom: 1px solid rgba(0, 0, 0, 0.12);
-        margin-bottom: 2rem;
-    }
-
-    .candidate-item {
-        display: flex;
-        margin: 2rem 0;
-        align-items: center;
-
-        .votes-count {
-            width: 6rem;
-            display: flex;
-            align-items: center;
-
-            .count {
-                background-color: rgba(0, 0, 0, 0.3);
-                padding: .4rem 1rem;
-                border-radius: .3rem;
-                color: #fff;
-            }
-        }
-
-        .candidate-name {
-            flex: 1;
-        }
-
-        .vote-btn {
-            margin-left: 2rem;
-            cursor: pointer;
-        }
-    }
-
-    .is-voted {
-        font-weight: bold;
-    }
-}
-
-.ant-modal-close {
-    box-shadow: none;
-    min-width: 0;
-
-    &:active {
-        box-shadow: none;
-    }
-}
-
-.ant-modal-title{
-    font-size: 23px;
-}
-
-.modal-header{
-    font-size:20px;
-}
-
-.modal-body {
-    margin: 2rem 0;
-    max-width: 32rem;
-
-    &__item {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin: 1rem 0;
-    }
-
-    &__category-list{
-        margin: 2rem 0;
-    }
-
-    &__label {
-        font-size:18px;
-        &:hover{
-            cursor: pointer;
-        }
-    }
-
-    &__checkbox {
-        width: 1.7rem;
-        height: 1.7rem;
-
-        &:hover{
-            cursor: pointer;
-        }
-    }
-}
